requirements.txt
asgiref==3.4.0
coverage==5.5
Django==3.2.4
django-extensions==3.1.3
djangorestframework==3.12.4
pygraphviz==1.7
pytz==2021.1
sqlparse==0.4.1
